requirements.txt
aiogram>=3.0.0
google-generativeai
Pillow
aiohttp-socks